changed cartTotal module
